Move BlueTriangleConfiguration into separate file
